move gui to separate file
